backend/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pypdf
numpy